chore: Split ActionCollection into two classes
