+++ app/server/appsmith-server/src/main/java/com/appsmith/server/domains/ActionCollection.java
@@ -1,9 +1,6 @@
 package com.appsmith.server.domains;
 
-import com.appsmith.external.models.BranchAwareDomain;
-import com.appsmith.external.views.Views;
-import com.appsmith.server.dtos.ActionCollectionDTO;
-import com.fasterxml.jackson.annotation.JsonView;
+import com.appsmith.server.domains.ce.ActionCollectionCE;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -19,38 +16,4 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @ToString
 @NoArgsConstructor
 @Document
-public class ActionCollection extends BranchAwareDomain {
-    // Default resources from BranchAwareDomain will be used to store branchName, defaultApplicationId and
-    // defaultActionCollectionId
-    @JsonView(Views.Public.class)
-    String applicationId;
-
-    // Organizations migrated to workspaces, kept the field as depricated to support the old migration
-    @Deprecated
-    @JsonView(Views.Public.class)
-    String organizationId;
-
-    @JsonView(Views.Public.class)
-    String workspaceId;
-
-    @JsonView(Views.Public.class)
-    ActionCollectionDTO unpublishedCollection;
-
-    @JsonView(Views.Public.class)
-    ActionCollectionDTO publishedCollection;
-
-    @Override
-    public void sanitiseToExportDBObject() {
-        this.setDefaultResources(null);
-        ActionCollectionDTO unpublishedCollection = this.getUnpublishedCollection();
-        if (unpublishedCollection != null) {
-            unpublishedCollection.sanitiseForExport();
-        }
-        ActionCollectionDTO publishedCollection = this.getPublishedCollection();
-        if (publishedCollection != null) {
-            publishedCollection.sanitiseForExport();
-        }
-        this.setOrganizationId(null);
-        super.sanitiseToExportDBObject();
-    }
-}
+public class ActionCollection extends ActionCollectionCE {}
