+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/ActionCollectionDTO.java
@@ -1,6 +1,7 @@
 package com.appsmith.server.dtos;
 
 import com.appsmith.server.dtos.ce.ActionCollectionCE_DTO;
+import com.querydsl.core.annotations.QueryEmbeddable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -10,4 +11,5 @@ import lombok.ToString;
 @Setter
 @NoArgsConstructor
 @ToString
+@QueryEmbeddable
 public class ActionCollectionDTO extends ActionCollectionCE_DTO {}
