+++ app/server/appsmith-server/src/main/java/com/appsmith/server/refactors/applications/RefactoringSolutionCE.java
@@ -5,9 +5,14 @@ import com.appsmith.server.dtos.LayoutDTO;
 import com.appsmith.server.dtos.RefactorEntityNameDTO;
 import reactor.core.publisher.Mono;
 
+import java.util.Set;
+
 public interface RefactoringSolutionCE {
 
     Mono<LayoutDTO> refactorEntityName(RefactorEntityNameDTO refactorEntityNameDTO, String branchName);
 
     Mono<Boolean> isNameAllowed(String contextId, CreatorContextType contextType, String layoutId, String newName);
+
+    Mono<Set<String>> getAllExistingEntitiesMono(
+            String contextId, CreatorContextType contextType, String layoutId, boolean isFQN);
 }
